refactor(pwa): move expense edit into route directory
Summary:
- Convert the expense edit file route into a route directory.
- Keep the edit history state helper beside the edit route.
- Commit the generated route tree update from the file-route move.

Rationale:
- The history marker is owned by the edit route behavior, so the helper should
  live in that route directory instead of the expense detail route directory.

Tests:
- vp run check
- vp exec playwright test e2e/expense-log.spec.ts -g "returns to (the expense log after editing|expense detail when backing out)"
