+++ packages/pwa/locale/en/messages.po
@@ -881,7 +881,7 @@ msgstr "Edit"
 msgid "Editing {0}"
 msgstr "Editing {0}"
 
-#: src/routes/party_.$partyId.expense.$expenseId_.edit.tsx:197
+#: src/routes/party_.$partyId.expense.$expenseId_.edit/route.tsx:197
 msgid "Editing {expenseName}"
 msgstr "Editing {expenseName}"
 
@@ -1014,7 +1014,7 @@ msgstr "Expense amounts don't match total. Please check your split configuration
 msgid "Expense Split Mode"
 msgstr "Expense Split Mode"
 
-#: src/routes/party_.$partyId.expense.$expenseId_.edit.tsx:161
+#: src/routes/party_.$partyId.expense.$expenseId_.edit/route.tsx:161
 msgid "Expense updated"
 msgstr "Expense updated"
 
@@ -1064,7 +1064,7 @@ msgstr "Failed to mark expense as paid"
 msgid "Failed to transfer debt"
 msgstr "Failed to transfer debt"
 
-#: src/routes/party_.$partyId.expense.$expenseId_.edit.tsx:166
+#: src/routes/party_.$partyId.expense.$expenseId_.edit/route.tsx:166
 msgid "Failed to update expense"
 msgstr "Failed to update expense"
 
@@ -2829,7 +2829,7 @@ msgstr "Update password"
 msgid "Update who you are in this party so that we can show you the expenses and stats that matter to you."
 msgstr "Update who you are in this party so that we can show you the expenses and stats that matter to you."
 
-#: src/routes/party_.$partyId.expense.$expenseId_.edit.tsx:130
+#: src/routes/party_.$partyId.expense.$expenseId_.edit/route.tsx:130
 msgid "Updating expense..."
 msgstr "Updating expense..."
